Export more things, the resumable sub scription creation in base client now looks quite nice actually
use existing method to get headers from string

Removed unused listeners

Retry strategy options

Forgot to commit this earlier.
